Clean up processing of received text.
Assert on unexpected conditions instead of commenting them as expected.
Use String.fromBTLE(utf8:) instead of using NSString (with the NSString.defaultCStringEncoding which is usually MacRoman, *not* UTF-8).
Don't mutate the UIARTTextView's textStorage on the background thread, since the mutation will post a notification that should only be sent on the main queue.